requirements.txt
numpy==1.24.4
pandas==2.1.4
Flask==3.0.0
python-dotenv==1.0.0
openpyxl==3.1.2
Werkzeug==3.0.1